+++ CHANGELOG.md
@@ -4,3 +4,10 @@
 
 ### Features
 - Initial commit of HA module
+
+## 0.0.3 (2025-05-3)
+
+### Features
+- Update to allow most recent version of Azure provider
+- Updating module to fix siteAddSecondaryAzureVSocket null resource
+- Added delay timer to allow HA configuration to complete